Add react menu bar without text
